Back to latest stable with directory restructure
