+++ package.json
@@ -7,7 +7,13 @@
     "test": "test"
   },
   "dependencies": {
-    "jsonfile": "^2.3.0"
+    "body-parser": "^1.15.2",
+    "cookie-parser": "^1.4.3",
+    "express": "^4.14.0",
+    "jsonfile": "^2.3.0",
+    "pug": "^2.0.0-beta3",
+    "request": "^2.72.0",
+    "spotify-web-api-node": "^2.3.0"
   },
   "devDependencies": {},
   "scripts": {
